feature: optimize constraint enumeration
Previously we enumerate over powerset of constraint, leading to overwhelming search space.
However, since this way doesn't leverage the property of equivalent relation.
For example, "a=b;b=c" actually equivalent to "a=b;b=c;a=c", enumerate both of them is useless.
Thus, now we enumerate all possible partitions of the set of symbols.